Open and close master.mdb on certain API responses
Closes #1.

We are inferring the API method from the structure of responses. This is not nice but otherwise we have to either hook stuff in il2cpp, or start a proxy server.

Does it introduce race conditions around new and delete? I think yes, but given that the game doesn't issue 2 requests simultaneously, I think we are mostly fine.